# docs/swagger.yaml
basePath: /
definitions:
...
  db.Transaction:
    properties:
      amount:
        type: number
      card_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      terminal_id:
        type: integer
    required:
    - amount
    type: object
...
  db.UpdateTransaction:
    properties:
      amount:
        type: number
      card_id:
        type: integer
      terminal_id:
        type: integer
    type: object
...
host: localhost:8888
info:
  contact: {}
  description: RESTful API server for MIFARE transactions authorization
  title: MIFARE API
  version: "0.1"
paths:
...
  /transactions:
    get:
      consumes:
      - application/json
      description: Возвращает список всех транзакций
      produces:
      - application/json
      responses:
        "200":
          description: Список транзакций
          schema:
            items:
              $ref: '#/definitions/db.Transaction'
            type: array
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получение всех транзакций
      tags:
      - transactions
    post:
      consumes:
      - application/json
      description: Создает новую транзакцию
      parameters:
      - description: Данные транзакции
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/db.Transaction'
      produces:
      - application/json
      responses:
        "200":
          description: ID транзакции
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Неверные параметры
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Нет доступа
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Создание транзакции
      tags:
      - transactions
  /transactions/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет транзакцию по ID
      parameters:
      - description: ID транзакции
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Успешно удалено
        "400":
          description: Неверный ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Нет доступа
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удаление транзакции
      tags:
      - transactions
    get:
      consumes:
      - application/json
      description: Возвращает данные транзакции по идентификатору
      parameters:
      - description: ID транзакции
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Данные транзакции
          schema:
            $ref: '#/definitions/db.Transaction'
        "400":
          description: Неверный ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получение транзакции по ID
      tags:
      - transactions
    put:
      consumes:
      - application/json
      description: Обновляет данные транзакции
      parameters:
      - description: ID транзакции
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для обновления
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/db.UpdateTransaction'
      produces:
      - application/json
      responses:
        "204":
          description: Успешно обновлено
        "400":
          description: Неверные параметры
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Нет доступа
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Обновление транзакции
      tags:
      - transactions
  /transactions/{id}/authorize:
    post:
      consumes:
      - application/json
      description: Пытается провести авторизацию транзакции по ID
      parameters:
      - description: ID транзакции
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Статус операции (approved/declined)
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверный ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Авторизация транзакции
      tags:
      - transactions
...
securityDefinitions:
  BearerAuth:
    description: 'Введите токен в формате: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"